Update design and images

+++ app/src/main/java/com/khnsoft/damta/RegisterArea6Fragment.kt
@@ -1,7 +1,9 @@
 package com.khnsoft.damta
 
 import android.app.Activity
+import android.app.AlertDialog
 import android.content.Intent
+import android.database.Cursor
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.graphics.Matrix
@@ -24,6 +26,7 @@ class RegisterArea6Fragment : Fragment() {
 
     companion object {
         const val RC_CAMERA = 100
+        const val RC_GALLARY = 101
         val ROOT_DIR = "${Environment.getExternalStorageDirectory()}/Hicarcom"
         val TEMP_DIR = "${ROOT_DIR}/.TEMP"
         val TEMP_NAME = "temp_image"
@@ -47,23 +50,38 @@ class RegisterArea6Fragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         btn_image.setOnClickListener {
-            val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
-            if (intent.resolveActivity(context!!.packageManager) != null) {
-                try {
-                    val dir = File(TEMP_DIR)
-                    dir.mkdirs()
+            AlertDialog.Builder(context)
+                .setTitle("이미지 추가")
+                .setItems(arrayOf("갤러리", "직접 촬영")) { dialog, which ->
+                    when (which) {
+                        0 -> {
+                            val intent = Intent(Intent.ACTION_PICK)
+                            intent.setType(MediaStore.Images.Media.CONTENT_TYPE)
+                            startActivityForResult(intent, RC_GALLARY)
+                        }
+                        1 -> {
+                            val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+                            if (intent.resolveActivity(context!!.packageManager) != null) {
+                                try {
+                                    val dir = File(TEMP_DIR)
+                                    dir.mkdirs()
 
-                    val file = File(TEMP_DIR, TEMP_NAME)
-                    if (file.exists()) {
-                        file.delete()
+                                    val file = File(TEMP_DIR, TEMP_NAME)
+                                    if (file.exists()) {
+                                        file.delete()
+                                    }
+
+                                    intent.putExtra(MediaStore.EXTRA_OUTPUT, FileProvider.getUriForFile(context!!, "${context?.packageName}.provider", file))
+                                    startActivityForResult(intent, RC_CAMERA)
+                                } catch (e: Exception) {
+                                    e.printStackTrace()
+                                }
+                            }
+                        }
                     }
-
-                    intent.putExtra(MediaStore.EXTRA_OUTPUT, FileProvider.getUriForFile(context!!, "${context?.packageName}.provider", file))
-                    startActivityForResult(intent, RC_CAMERA)
-                } catch (e: Exception) {
-                    e.printStackTrace()
                 }
-            }
+                .create()
+                .show()
         }
     }
 
@@ -86,6 +104,29 @@ class RegisterArea6Fragment : Fragment() {
                         e.printStackTrace()
                     }
                 }
+            }
+            RC_GALLARY -> {
+                if (data == null) return
+                val photoUri = data.data
+                var cursor: Cursor? = null
+                val tempFile: File?
+
+                try {
+                    val proj = arrayOf(MediaStore.Images.Media.DATA)
+
+                    cursor = context!!.contentResolver.query(photoUri ?: return, proj, null, null, null)
+
+                    val column_index = cursor?.getColumnIndexOrThrow(MediaStore.Images.Media.DATA) ?: return
+
+                    cursor.moveToFirst()
+                    tempFile = File(cursor.getString(column_index))
+                } finally {
+                    cursor?.close()
+                }
+
+                val options = BitmapFactory.Options()
+                val img = BitmapFactory.decodeFile(tempFile?.absolutePath ?: return, options)
+                showImage(img)
             }
         }
     }
